Changing the parameter data type of the method getByNumber in the IRoomRepository from the domain/port directory

+++ src/main/java/dev/waamir/hotelaluraapi/domain/port/IRoomRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface IRoomRepository<T extends Room> {
     T create(T t);
     Optional<T> getById(Long id);
-    Optional<T> getByNumber(int number);
+    Optional<T> getByNumber(Long number);
     void delete(T t);
     void update(T t);
     List<T> list();
